renamed the namespace Microsoft.VisualStudio to Mono.VisualStudio so this can build in a vsix solution
