+++ Mono.TextTemplating/Mono.TextTemplating/ParsedTemplate.cs
@@ -28,7 +28,7 @@ using System;
 using System.CodeDom.Compiler;
 using System.Collections.Generic;
 using System.IO;
-using Microsoft.VisualStudio.TextTemplating;
+using Mono.VisualStudio.TextTemplating;
 
 namespace Mono.TextTemplating
 {
