+++ Mono.TextTemplating/Mono.TextTemplating/TemplateGenerator.cs
@@ -30,7 +30,7 @@ using System.CodeDom.Compiler;
 using System.IO;
 using System.Reflection;
 using System.Text;
-using Microsoft.VisualStudio.TextTemplating;
+using Mono.VisualStudio.TextTemplating;
 
 namespace Mono.TextTemplating
 {
