+++ Mono.TextTemplating/Mono.TextTemplating/TemplateSettings.cs
@@ -27,7 +27,7 @@
 using System;
 using System.Text;
 using System.Collections.Generic;
-using Microsoft.VisualStudio.TextTemplating;
+using Mono.VisualStudio.TextTemplating;
 using System.IO;
 
 namespace Mono.TextTemplating
