+++ Mono.TextTemplating/Mono.VisualStudio.TextTemplating/TextTransformation.cs
@@ -29,7 +29,7 @@ using System.CodeDom.Compiler;
 using System.Collections.Generic;
 using System.Text;
 
-namespace Microsoft.VisualStudio.TextTemplating
+namespace Mono.VisualStudio.TextTemplating
 {
 	public abstract class TextTransformation : IDisposable
 	{
